Made it clear this is for 2D displays
Also made it clear that the technique is made in such a way that it allows interventions at different points in the technique depending on the content. The additional control given developers results in a technique which uses only linear transformations - more complex transformations wouldn't allow for similar interventions. 
(Issue #36)